Move object parsing to `syntax/object`
