+++ crates/rslint_parser/src/syntax/expr.rs
@@ -5,12 +5,12 @@
 
 use syntax::decl::is_semi;
 
-use super::decl::{arrow_body, class_decl, maybe_private_name, method, parameter_list};
+use super::decl::{arrow_body, class_decl, maybe_private_name, parameter_list};
 use super::pat::pattern;
 use super::typescript::*;
 use super::util::*;
-use crate::syntax::decl::formal_param_pat;
-use crate::syntax::function::{function_body, function_expression, ts_return_type};
+use crate::syntax::function::function_expression;
+use crate::syntax::object::object_expr;
 use crate::{SyntaxKind::*, *};
 
 pub const EXPR_RECOVERY_SET: TokenSet = token_set![VAR_KW, R_PAREN, L_PAREN, L_BRACK, R_BRACK];
@@ -1156,234 +1156,6 @@ pub fn spread_element(p: &mut Parser) -> CompletedMarker {
 	p.expect(T![...]);
 	assign_expr(p);
 	m.complete(p, SPREAD_ELEMENT)
-}
-
-/// An object literal such as `{ a: b, "b": 5 + 5 }`.
-// test object_expr
-// let a = {};
-// let b = {foo,}
-pub fn object_expr(p: &mut Parser) -> CompletedMarker {
-	let m = p.start();
-	p.expect(T!['{']);
-	let props_list = p.start();
-	let mut first = true;
-
-	while !p.at(EOF) && !p.at(T!['}']) {
-		if first {
-			first = false;
-		} else {
-			p.expect(T![,]);
-			if p.at(T!['}']) {
-				break;
-			}
-		}
-		object_member(p);
-	}
-	props_list.complete(p, LIST);
-
-	p.expect(T!['}']);
-	m.complete(p, JS_OBJECT_EXPRESSION)
-}
-
-const STARTS_OBJ_PROP: TokenSet = token_set![
-	JS_STRING_LITERAL_TOKEN,
-	JS_NUMBER_LITERAL_TOKEN,
-	T![ident],
-	T![await],
-	T![yield],
-	T!['[']
-];
-
-/// An individual object property such as `"a": b` or `5: 6 + 6`.
-pub fn object_member(p: &mut Parser) -> Option<CompletedMarker> {
-	let m = p.start();
-
-	match p.cur() {
-		// test object_expr_getter_setter
-		// let a = {
-		//  get foo() {
-		//    return foo;
-		//  }
-		// }
-		T![ident]
-			if p.cur_src() == "get"
-				&& !p.has_linebreak_before_n(1)
-				&& STARTS_OBJ_PROP.contains(p.nth(1)) =>
-		{
-			Some(getter_object_member(p))
-		}
-
-		// test object_expr_getter_setter
-		// let b = {
-		//  set [foo](bar) {
-		//     return 5;
-		//  }
-		// }
-		T![ident]
-			if p.cur_src() == "set"
-				&& !p.has_linebreak_before_n(1)
-				&& STARTS_OBJ_PROP.contains(p.nth(1)) =>
-		{
-			Some(setter_object_member(p))
-		}
-
-		// test object_expr_async_method
-		// let a = {
-		//   async foo() {},
-		//   async *foo() {}
-		// }
-		T![ident]
-			if p.cur_src() == "async"
-				&& !p.has_linebreak_before_n(1)
-				&& (STARTS_OBJ_PROP.contains(p.nth(1)) || p.nth_at(1, T![*])) =>
-		{
-			method(p, None, None)
-		}
-
-		// test object_expr_spread_prop
-		// let a = {...foo}
-		T![...] => {
-			p.bump_any();
-			assign_expr(p);
-			Some(m.complete(p, SPREAD_PROP))
-		}
-
-		T![*] => {
-			// test object_expr_generator_method
-			// let b = { *foo() {} }
-			let m = p.start();
-			method(p, m, None)
-		}
-
-		_ => {
-			let member_name = object_member_name(p);
-
-			// test object_expr_method
-			// let b = {
-			//  foo() {},
-			// }
-			if p.at(T!['(']) || p.at(T![<]) {
-				if let Some(mut member_name) = member_name {
-					member_name.change_kind(p, NAME);
-				}
-
-				method(p, m, None)
-			} else if let Some(mut member_name) = member_name {
-				// test object_expr_assign_prop
-				// let b = { foo = 4, foo = bar }
-				if p.eat(T![=]) {
-					member_name.change_kind(p, NAME);
-					assign_expr(p);
-					return Some(m.complete(p, INITIALIZED_PROP));
-				}
-
-				// ({foo})
-				// test object_expr_ident_prop
-				if !p.at(T![:]) && member_name.kind() == JS_STATIC_OBJECT_MEMBER_NAME {
-					member_name.change_kind(p, JS_REFERENCE_IDENTIFIER_EXPRESSION);
-					Some(m.complete(p, JS_SHORTHAND_PROPERTY_OBJECT_MEMBER))
-				} else {
-					// let b = { a: true }
-					// If the member name was a literal OR we're at a colon
-					p.expect(T![:]);
-					assign_expr(p);
-					Some(m.complete(p, JS_PROPERTY_OBJECT_MEMBER))
-				}
-			} else {
-				// test_err object_expr_error_prop_name
-				// let a = { /: 6, /: /foo/ }
-				// let a = {{}}
-				// test_err object_expr_non_ident_literal_prop
-				// let b = {5}
-
-				p.err_recover_no_err(token_set![T![:], T![,]], false);
-
-				if p.eat(T![:]) {
-					assign_expr(p);
-					Some(m.complete(p, JS_PROPERTY_OBJECT_MEMBER))
-				} else {
-					None
-				}
-			}
-		}
-	}
-}
-
-fn getter_object_member(p: &mut Parser) -> CompletedMarker {
-	debug_assert!(p.at(T![ident]), "Expected an identifier");
-	debug_assert!(p.cur_src() == "get", "Expected a get identifier");
-
-	let m = p.start();
-
-	p.bump_remap(T![get]);
-
-	object_member_name(p);
-
-	p.expect(T!['(']);
-	p.expect(T![')']);
-
-	ts_return_type(p);
-
-	function_body(p);
-
-	m.complete(p, JS_GETTER_OBJECT_MEMBER)
-}
-
-fn setter_object_member(p: &mut Parser) -> CompletedMarker {
-	debug_assert!(p.at(T![ident]), "Expected an identifier");
-	debug_assert!(p.cur_src() == "set", "Expected a set identifier");
-
-	let m = p.start();
-
-	p.bump_remap(T![set]);
-
-	object_member_name(p);
-
-	p.state.allow_object_expr = p.expect(T!['(']);
-	formal_param_pat(p);
-	p.expect(T![')']);
-
-	function_body(p);
-
-	p.state.allow_object_expr = true;
-	m.complete(p, JS_SETTER_OBJECT_MEMBER)
-}
-
-// test object_prop_name
-// let a = {"foo": foo, [6 + 6]: foo, bar: foo, 7: foo}
-pub fn object_prop_name(p: &mut Parser, binding: bool) -> Option<CompletedMarker> {
-	match p.cur() {
-		JS_STRING_LITERAL_TOKEN | JS_NUMBER_LITERAL_TOKEN => literal(p),
-		T!['['] => {
-			let m = p.start();
-			p.bump_any();
-			assign_expr(p);
-			p.expect(T![']']);
-			Some(m.complete(p, COMPUTED_PROPERTY_NAME))
-		}
-		_ if binding => super::pat::binding_identifier(p),
-		_ => identifier_name(p),
-	}
-}
-
-// test object_prop_name
-// let a = {"foo": foo, [6 + 6]: foo, bar: foo, 7: foo}
-/// Parses a `JsAnyObjectMemberName` and returns its completion marker
-fn object_member_name(p: &mut Parser) -> Option<CompletedMarker> {
-	match p.cur() {
-		JS_STRING_LITERAL_TOKEN | JS_NUMBER_LITERAL_TOKEN => literal(p),
-		T!['['] => {
-			let m = p.start();
-			p.bump_any();
-			assign_expr(p);
-			p.expect(T![']']);
-			Some(m.complete(p, COMPUTED_PROPERTY_NAME))
-		}
-		_ => identifier_name(p).map(|mut ident| {
-			ident.change_kind(p, JS_STATIC_OBJECT_MEMBER_NAME);
-			ident
-		}),
-	}
 }
 
 /// A left hand side expression, either a member expression or a call expression such as `foo()`.
